chore: Update AuthService to include getCurrentUserState method
The AuthService has been updated to include a new method called getCurrentUserState. This method returns an Observable of type UserState, which represents the current state of the user. The method uses the onAuthStateChanged method from the firebaseAuth service to listen for changes in the user's authentication state. If a user is authenticated, the method also checks if the user is an admin by calling the isUserAdmin method from the UsersFbService. If the user is not authenticated, the method returns an object with user set to null and isAdmin set to false. This update improves the functionality of the AuthService by providing a more comprehensive representation of the user's state.