chore: Update login component to include loginType based on router URL
The login component has been updated to include the loginType property based on the router URL. This allows the component to determine the type of login (email or phone) based on the current route. By setting the loginType property, the component can customize its behavior accordingly. This update improves the functionality of the login component by dynamically adapting to the login type specified in the URL.