refactor: Improve delete functionality in CitafuturaComponent
Refactor the `btnEliminarCita` method in the `CitafuturaComponent` to improve the delete functionality. Instead of immediately deleting the cita, a confirmation dialog is shown using the `Swal.fire` library. If the user confirms the deletion, the cita is deleted using the `CitasFbService.deleteCita` method. After the deletion, a success or error message is displayed using `Swal.fire`. This refactor improves the user experience and provides better feedback when deleting a cita.